+++ src/pro/pro_rtp/rtp_msg_c2s.cpp
@@ -753,7 +753,8 @@ CRtpMsgC2s::AcceptSession(IRtpService*            service,
         CProStlString theString = "";
         msgStream.ToString(theString);
 
-        if (!m_msgClient->SendMsg(theString.c_str(), theString.length(), 0, &ROOT_ID_C2S, 1))
+        if (!m_msgClient->SendMsg(theString.c_str(), (unsigned long)theString.length(),
+            0, &ROOT_ID_C2S, 1))
         {
             m_reactor->CancelTimer(timerId);
 
@@ -1584,5 +1585,6 @@ CRtpMsgC2s::ReportLogout(IRtpMsgClient*      msgClient,
     CProStlString theString = "";
     msgStream.ToString(theString);
 
-    msgClient->SendMsg(theString.c_str(), theString.length(), 0, &ROOT_ID_C2S, 1);
+    msgClient->SendMsg(theString.c_str(), (unsigned long)theString.length(),
+        0, &ROOT_ID_C2S, 1);
 }
