+++ src/pro/test_msg_client/test.cpp
@@ -292,7 +292,7 @@ CTest::SendMsg(const char* msg)
             dstUsers.push_back(*itr);
         }
 
-        m_msgClient->SendMsg(msg, strlen(msg),
+        m_msgClient->SendMsg(msg, (unsigned long)strlen(msg),
             0, &dstUsers[0], (unsigned char)dstUsers.size());
     }
 }
@@ -335,7 +335,7 @@ CTest::SendMsg(const char*         msg,
             return;
         }
 
-        m_msgClient->SendMsg(msg, strlen(msg), 0, dstUser, 1);
+        m_msgClient->SendMsg(msg, (unsigned long)strlen(msg), 0, dstUser, 1);
     }
 }
